Reorganized into MVVM architecture
